refactor: productLike DTO 패키지 변경
