+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/product/ProductQueryDslRepository.java
@@ -5,10 +5,8 @@ import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.loopers.infrastructure.product.ProductQueryDslRepositoryImpl.ProductWithLikeCount;
-
 public interface ProductQueryDslRepository {
     Page<ProductWithLikeCount> findByBrandAndLikeCount(String sortKey, Long brandId, Pageable pageable);
 
-    Page<ProductWithLikeCount> findByBrandAndLikeCountDenormalization(String sortKey, List<Long> brandId, Pageable pageable);
+    Page<ProductWithLikeCount> findByBrandDenormalization(String sortKey, List<Long> brandId, Pageable pageable);
 }
