+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/product/ProductWithLikeCount.java
@@ -1,0 +1,7 @@
+package com.loopers.infrastructure.product;
+
+import com.loopers.domain.brand.Brand;
+import com.loopers.domain.product.Product;
+
+public record ProductWithLikeCount(Product product, Brand brand, Long likeCount) {
+}
